Feature: mchan logging (#20)
* Move *request* into separate ns

In preparation for refactoring which would otherwise cause cyclic
dependency.

* Introduce context parameter to put! signature

Then pass it on to useful.channel/publish!

* Propagate context to more deeply nested hooks

* Source useful lib with updated put! implementation

* Add version number under 1.6.7

With an eye on backporting changes to master soon.

* Generate a context if none is found

Provides context for p/conj!, p/update-in! and p/destroy! calls which
happen neither in a web handler nor in an mchan hander.

* Source latest version of useful lib

* Bump junior patch version

* Bump junior patch version

* Source latest useful lib

* Reduce visibility

* Replace f-wrapper with Hook implementation

* Bump patch version

* Bump major version

* Use context abstraction feature

Wrap the current request in an object which implements the Context
protocol.

* Downgrade validateur

Version 2.6.0 breaks our tests. Our source needs adaptation before we
can upgrade validateur.

* Remove logging dependencies

The domain library logs nothing.

* Source latest domain version

* domain_types: Bump patch version

* domain_types: Upgrade clojure and useful

* Upgrade congomongo

* Migrate domain-dependent code

This used to be in useful, causing a circular dependency. Bringing it
into domain breaks that dependency and puts the code into its proper,
intuitive place.

* Remove debugging artifact